Move TestStagingCerts to its own file
TestStagingCerts covers both Compose (Caddyfile, rewrite_caddy) and
Kubernetes (values.yaml, cert-manager, side-effect propagation), so
living in test_kubernetes_structure.py was misleading.

Move the class and its Jinja render helpers (_ansible_jinja_env,
_render_template) to a new tests/unit/test_staging_certs.py. Test
coverage unchanged (18 cases); just organized by feature instead of
by orchestrator.
